refactor: move search factory
